kvm-xfstests, test-appliance: enhance host<->client protocol
The protocol between the host and the client is now versioned.  In
addition, we now pass an set of test options and the primary file
system type (which isn't used yet, but which will allow us to provide
first-class for file systems beyond ext4).

Add the -C option to kvm-xfstests, which allows us to run a particular
test multiple times.  This is useful for flaky tests.

Signed-off-by: "Theodore Ts'o" <tytso@mit.edu>