test-appliance: use e2fsck -E discard
In compress-rootfs, use e2fsck -E discard instead of the zerofree
binary.  It's faster, since we don't have to write zeros into the raw
file system, but instead take advantage of the punch hole functionality.

Also use this when creating the root_fs image so we don't have to
manually run compress-rootfs after creating the image.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>