Include -dirty in git versions
This is a good safety check in case there are some local changes that
haven't been checked in yet.

Also move the creation of xfstests-bld.ver to the "make tarball"
target.  That way we can make changes to the run time files and
rebuild the tarball without having to rerun the entire build.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>