test-appliance: add create-local-mirror config script
This script mines the cached files in var.cache.apt.archives and
var.lib.apt.lists to create a (subset) debian mirror.  This allows us
to run the gen-image script without having access to the network.

For example, after running the commands:

./gen-image --mirror http://snapshot.debian.org/archive/debian/20141013T184415Z
./create-local-mirror snapshot.debian.org_archive_debian_20141013T184415Z

It becomes possible to execute the command:

./gen-image --mirror file:mirror

Which avoids needing to hit snapshot.debian.org each time we run
the gen-image script.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>