test-appliance: add --update option to the gen-image script
This allows gen-image to to update the xfstests.tar.gz tarball and the
xfstests driver scripts into an existing root_fs.img file, keeping the
original operating systems files unchanged.  This takes less than 15
seconds, compared to about 90 seconds to do a full gen-image build
using a local disk debian mirror.  It also allows a xfstests developer
to generate new versions of the root_fs.img more simply on a system
that does not have debootstrap installed or has access to a Debian
archive or mirror.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>