Use a specific commit for fio and quota repositories
The fio and quota repositories are not always stable and sometimes
don't build correctly and/or have bugs.  In the case of the quota
repository, the latest versions have been autoconfiscated, and the
build system hasn't entirely stable.

So use the last officially released versions of fio and quota so that
other people who check out xfstests-bld don't get confused.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>