test-appliance: modify the ext3 config to disable all non-ext4 features
Now that e2fsprogs 1.43.x is enabling the 64bit feature by default for
ext4 file systems, we need to disable it explicitly in the ext3
config.  While we're at it, disable all of the ext4 features so the
ext3 config tests what would be created using mke2fs -t ext3.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>