xfstests-bld: exclude ext4 defrag tests from unsupported configurations
Online defrag is not currently implemented for journaled data and
bigalloc ext4 file systems.  The xfstests that exercise online defrag
do not verify that the test file system supports that feature before
executing, resulting in spurious test failures.  (In the case of
ext4/307, the test reports success but actually fails because the
e4compact test component is ignoring failures.)

For the time being, simply exclude these tests to eliminate the test
failure noise.

Signed-off-by: Eric Whitney <enwlinux@gmail.com>
Signed-off-by: Theodore Ts'o <tytso@mit.edu>