add update-all script and change do-update so it calls this script
Add an update-all script which rebuilds all of the components (but
without rerunning the configure step, and which recompiles all of the
.ver files.

This means that the primary difference between do-update and do-all is
that do-update skips running configure scripts (which can be quite
time-consuming) and it updates the xfstests tarball in place without
regenerating the debian image.  On my system the do-update script
takes about 50 seconds, while the do-all script takes a little over 2
minutes (and requires network access).

Signed-off-by: Theodore Ts'o <tytso@mit.edu>