build: pass cross compilation options to configure
Add support for CROSS_COMPILE and TOOLCHAIN_DIR in the config file and
if cross compilation has been requested, pass the appropriate --build
and --host options to the configure scripts.

Not all of the components built in xfstests-bld support
cross-compilation correctly, or have glibc dependencies that will
break when built for other runtime environments (such as Android).  To
be fixed in later commits (hopefully).

Signed-off-by: Theodore Ts'o <tytso@mit.edu>