gce-xfstests, kvm-xfstests: improve command line parsing
Print an error message if an invalid test configuration is specified
or if no tests are specified.

Add the --no-action option for debugging

Customize the help message for gce-xfstests and disable command and
options that are not applicable for gce-xfstests (e.g., "maint",
"--no-log", "--log", "-N").

Add a default memory configuration for GCE to prevent OOM errors
caused by the background ruby monitoring script.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>