gce-xfstests: teach gce-xfstests the maint subcommand
The maint command will start an f1-micro instance and bypass kexec'ing
into a user-supplied kernel.  This allows the administrator to attach
failed results disks, or other things which are easier with shell in
GCE.

Also add some new options which were used in making "maint" work.
Added the --machtype option which allows the user to specify a
specific instance machine type.  Also added support for "--kernel
none" which will skip running kexec to switch to a user-specified
kernel.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>