test-appliance: run gce-setup for "gce-xfststs shell"
The use of "gce-xfststs shell", which kexec's into a test kernel, is
almost always to run tests.  So fire up /usr/local/lib/gce-setup as
part of the boot process to save time.  We will not do this for
"gce-xfstests maint", which uses the default Debian kernel and does
not do a kexec, since there's no point wasting resources.  We will
mark the results volume for automatic delection since we generally
don't want to save it, and it will get in the way of the next
"gce-xfstests shell" command.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>