test-appliance: don't update the saved check-time after each repeated test
If the user has specified that a test (or tests) be repeated multiple
times, don't save the check-time.tar.gz file back to Google Cloud
Storage after each test rep.  Instead, only save it when we are done
running that particular test configuration.  This reduces the overhead
when running a potentially flaky test multiple times.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>