misc: fix encrypt-fname-benchmark to use a 4k blocksize
The kernel now enforces that file systems using encryption must have
their block size to be the same as the page size.  It doesn't matter
for file name encryption, which is why we didn't notice that we were
using a 1k block size when testing the filename encryption/decryption
speeds.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>