test-appliance: add new gce-finalize service to handle clean test completions
The original scheme of handling gce-shutdown was failing on clean
shutdowns due to gce-shutdown trying to stop the kvm-xfstests service
--- which ended up killing itself if it was executed from the
kvm-xfstests.boot shell script.

So now we have a completely separate service for finalizing the
shutdown in the clean completion case which is gated by the file
/run/gce-finalize-wait.  When the file is deleted, then the
finalization/shutdown process proceeds.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>