gce-xfstests: support multiple extra kernel options
We need to escape the space when passing it up to the GCE VM.  We'll
use the colon (':') character for this purposes since it's not really
used in any of the boot command line options we're likely to care
about.

This requires a bump of the test-runner/VM interface version number;
we are now using kvm-xfstests API version 1.3.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>