Use LIBS instead of LDFLAGS in build of dbench
The Makefile of dbench doesn't really honor LDFLAGS. It's a really
broken Makefile overall and hasn't been fixed upstream. So, for now,
let's work around it by passing those flags through LIBS instead.

Also don't link it with -static since that doesn't really work and
produces the following warning:
  Using 'gethostbyname' in statically linked applications requires at
  runtime the shared libraries from the glibc version used for linking

Tested that it links against the static libpopt.a in bld/lib, even when
there's a shared library version of libpopt in the build system.

Signed-off-by: Filipe Brandenburger <filbranden@google.com>
Signed-off-by: Theodore Ts'o <tytso@mit.edu>