get-all: avoid regenerating the generated files in acl/ during build
Due to the vagrancies of the timestamps when this repository is
checked out, the Makefile might try to regenerate the aclocal.m4,
configure, and/or Makefile.in files in the acl directory.  If the the
dependencies for creating these generated files are not be available,
this will cause build failures.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>