kvm-xfstests,gce-xfstests: move logic to read config.custom out of config
We are inconsistent about whether it is safe to copy config to
config.custom; we tell people who want to build xfstests.tar.gz to do
this, but it's unsafe to do this for kvm-xfstests, since it will cause
config.custom to try to source itself.  To avoid this problem, move
the logic of deciding which config file(s) to source to the
util/get-config shell script fragment.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>