Add a Dockerfile for kvm-xfstests
Unlike the top-level Dockerfile, this one is designed to create a
small, optimized Docker image that most users of kvm-xfststs will want
to use.  It does not try to build xfstests or the root_fs.img;
instead, it will use the prebuilt root_fs.img in the local
test-appliance directory, or a prebuilt root_fs.img uploaded
www.kernel.org.

Also move the kvm-xfstests.in and gce-xfstests.in from the top-level
directory to the kvm-xfstests directory.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>