kvm-xfstests: forbid conflicting specifications of FSTESTCFG
It was possible to run kvm-xfstests with arguments like '-c encrypt
smoke', in which case FSTESTCFG ended up set to "4k" because 'smoke'
implies '-c 4k' and the second config specification overrode the first.
If the user tries something like this, fail with a helpful error message
rather than proceeding with a configuration that likely wasn't intended.

Signed-off-by: Eric Biggers <ebiggers@google.com>
Signed-off-by: Theodore Ts'o <tytso@mit.edu>