gce-xfstests: when building image don't leave git-versions in source tree
We update git-versions and put it in files/root/xfstests/git-versions
just case there have been changes made to the test appliance test
runner scripts that should be reflected in the xfstets-bld git
version.

However, if we leave this file around, it may interfere with the
git-versions file uploaded via gce-xfstests --update-xfstests.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>