xfstests-bld: correct file permissions on test appliance files
The xfstests-bld repository may have been cloned with a umask that
masked out the group and/or the other bits.  When using --update-files
in this situation or when creating a GCE image, the VM ended up in a
state where non-root users were unable to execute the binaries in
/root/xfstests, which made all tests using the fsgqa user get skipped.
Fix this by copying the r and x user bits to the group and other bits
when creating files.tar.gz, creating xfstests.tar.gz, or copying files
directly to the kvm-xfstests appliance.  Also consistently set the owner
and group to root.

Note that git doesn't actually store file owners, groups, or any mode
bits other than "is the file executable"?  So this patch really just
enforces a consistent, sane default for this "unstored" metadata.

Signed-off-by: Eric Biggers <ebiggers@google.com>
Signed-off-by: Theodore Ts'o <tytso@mit.edu>