test-appliance: use old key prefix in ext4 encrypt config
For the ext4 encrypt config, add the dummy key using the "ext4:" prefix
instead of the "fscrypt:" prefix.  This will allow xfstests-bld to test
any old kernels to which the change to make test_dummy_encryption
require a keyring key may get backported.  (Of course, it also remains
able to test old kernels to which that change has *not* been
backported.)  Using the old / fs-specific key prefix is also consistent
with the xfstests in the "encrypt" group, which use the fs-specific key
prefix to allow the tests to run unmodified on older kernels.

Signed-off-by: Eric Biggers <ebiggers@google.com>
Signed-off-by: Theodore Ts'o <tytso@mit.edu>