remove kernel-configs for 4.2
I am not aware of anyone who is maintaining 4.2 as a stable version, and
it's just one version ahead of 4.1 which actually is being maintained as
an upstream stable version.  So it seems it isn't worthwhile to maintain
kernel configs for 4.2 in xfstests-bld anymore, and in the off-chance
someone wants them they can just base their config on 4.1.

Signed-off-by: Eric Biggers <ebiggers@google.com>
Signed-off-by: Theodore Ts'o <tytso@mit.edu>