kvm-xfstests: add support for using virtfs
Optionally, if guest kernel is compiled with 9P support, the file
system developer will be able to mount /vtmp, and then be able to read
and write into /tmp/kvm-xfstests-$USER on the host operating system.

We deliberately do not give the guest kernel free run of host file
system, as this could cause some corporate security teams to have
kittens.  :-)

Signed-off-by: Theodore Ts'o <tytso@mit.edu>