test-appliance: use the large scratch device when the config size is large
There was a bug which caused the use of the small scratch device even
when use of large devices was requested in the test scenario file.
This was cuased by some legacy default code that should have been
removed when we added the support for large devices as an abstract
request.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>