test-appliance: yet another file permissions fix
Several directories in the kvm-xfstests appliance were still not being
given the correct file permissions, since they were rsync'ed directly
into the mounted root filesystem with the '-a' option.  Unfortunately it
seems we can't express "go+u-w" with rsync options (it's not accepted by
--chmod), but we can copy the files through 'tar' instead, similar to
what is done elsewhere.

(I didn't notice this earlier because it is kvm-xfstests specific, and
it only broke a couple tests such as generic/128.)

Signed-off-by: Eric Biggers <ebiggers@google.com>
Signed-off-by: Theodore Ts'o <tytso@mit.edu>