test-appliance: checkpoint fs configurations that are completed
If the test VM is stopped (for example, by preemption), when it is
restarted, skip those configurations which have been fully finished.
We'll rerun all of the tests in a fs configuration pass, since
sometimes a bug is triggered by the tests being run in sequence.  It's
not ideal this is true, but it's also simpler to checkpoint on a per
file system configuration basis, so we'll make a virtue of necessity
(or at least, a simpler implementation plan).

Signed-off-by: Theodore Ts'o <tytso@mit.edu>