android-xfstests: support userdata on device-mapper device
Update android-xfstests to support setting up the xfstests partitions on
Android devices where the userdata filesystem is located on a
device-mapper device backed by the userdata partition, rather than on
the userdata partition itself.  This is the case for devices using
"full-disk encryption", for example, in which case dm-crypt is used.

We can do things pretty similarly to the non-dm case: just shrink the
raw userdata partition after reformatting the filesystem to be smaller,
then create the test partitions in the newly "free" space.  Shrinking
the dm device itself is not required.  We do however need to make sure
not to clobber any footer, e.g. a crypto footer, that may be present.

Note that it's not possible to create the test partitions on top of the
dm device.  However, it would be possible to add an option to set up
separate dm devices for the test partitions.

Signed-off-by: Eric Biggers <ebiggers@google.com>
Signed-off-by: Theodore Ts'o <tytso@mit.edu>