test-appliance: get rid of cruft output on serial ports
Don't try to run the resize command on the serial ports, since there
won't be anything connected to the serial ports when the VM is started.

Also transfer the responsibility of creating the /dev/vd[bcdef]
convenient symlinks to gce-xfstests-bld.sh, to avoid races from
multiple shells starting up on the various serial ports.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>