test-appliance: revert logging change on startup
Commit 458ce3391003 ("Add support to optionally build gce-xfstests for
Debian stretch") made the output of kvm-xfstests.boot get redirected to
a file /var/log/kvm-xfstests-boot.log.  This actually isn't desirable,
because the output is supposed to be shown on the console.  This was
especially apparent for kvm-xfstests, since it hid the output of
runtests.sh.  But it was also applicable to gce-xfstests, since
'gce-xfstests console' can be used to show the console output.

We could send the output to both places using 'tee', but this seems
useless since the output can just be retrieved from the console instead
of from the log file.  Also, the 'set -vx' output is too verbose for the
console, and I don't think there would be a way to send it just to the
log file.

So just revert the change to kvm-xfstests.boot.

Signed-off-by: Eric Biggers <ebiggers@google.com>
Signed-off-by: Theodore Ts'o <tytso@mit.edu>