get-all: check out correct commits in already-cloned repositories
Update the get-all script to check out the requested commit in each git
repository, even if the repository has already been cloned.  Previously,
the repositories would become outdated if the config file was changed.

It was also possible for local changes (even uncommitted changes) to be
included in the build --- maybe intentionally, maybe unintentionally.
But there is a clear way to express the difference now: just comment out
${REPO_NAME}_COMMIT in config.custom if it's desired to manage a
repository manually and do the builds from whatever happens to be in the
working tree rather than from a specific commit.

Signed-off-by: Eric Biggers <ebiggers@google.com>
Signed-off-by: Theodore Ts'o <tytso@mit.edu>