xfstests-bld: optionally build keyctl
Add support for optionally including the keyctl program in the xfstests
tarball built by xfstests-bld.  keyctl is used by the filesystem
encryption tests.  Although keyctl is already included in the test
appliances, including it in the xfstests tarball can be useful in cases
where the tarball is being used on its own, without being contained in a
dedicated test appliance.

Signed-off-by: Eric Biggers <ebiggers@google.com>
Signed-off-by: Theodore Ts'o <tytso@mit.edu>