kernel-configs: remove configs for non-LTS versions
Remove configs for the 3.3, 3.6, 3.8, and 4.7 kernels, as those kernels
were not maintained for very long so are not as likely to be tested.

This was mostly an issue for the i386 configs, which were sometimes for
different versions than the x86_64 configs.  Therefore, this patch
removes several i386 configs.  This will be made up for by ensuring we
have a corresponding i386 config for each x86_64 config.

Signed-off-by: Eric Biggers <ebiggers@google.com>
Signed-off-by: Theodore Ts'o <tytso@mit.edu>