kernel-configs: sync i386 configs with x86_64 ones
Now that the x86_64 configs have been updated, update/create the
corresponding i386 configs.  For modern kernels this just meant
explicitly unsetting CONFIG_64BIT in each config and doing 'make
olddefconfig; make savedefconfig'.  3.2 and 3.4 are different; to
work with those, ARCH=i386 must also be set in the environment.

Note that the i386 configs only work with kvm-xfstests, since the
userspace in the GCE appliances is 64-bit.

Signed-off-by: Eric Biggers <ebiggers@google.com>
Signed-off-by: Theodore Ts'o <tytso@mit.edu>