gce-xfstests: qualify instance names with username and seconds
The DATECODE used in running gce-xfstests now uses seconds granularity rather
than just minutes. Additionally, by default, all gce-xfstests instances will
be launched with "$USER-" prepended to the datecode, for better
identification of test runs/results. This can be overridden in
~/.config/gce-xfstests using GCE_USER=<non-empty string> or even set to
the empty string ("GCE_USER=") to disable the USER string prepend entirely.
As this is not a DATECODE as much as it is a "test run identifier", the
variable name has been changed to TESTRUNID, or "testrunid" for the
filename in the /results directory on the test appliance.

This is to help distinguish instances created by multiple users in the same
gce-xfstests project. The empty string flag is intended to allow single user
projects to leave out the username.

Tested: launched test appliance, test appliance name had seconds and
username in it.

Signed-off-by: Jason Liu <liujason@google.com>
Change-Id: I569c85986973ec16f11e990704228328b74c7162