test-appliance: disable lazy_itable_init for lustre_mds
This is to work around an apparent bug GCE's PD implementation.  It
looks like issueing a large zeroout (when using a 2048 byte inode)
during lazy inode table initialization is making the persistent disk
implementation very unhappy.  This causes the test VM to crash, so
until this can be investigated and hopefully fixed, initializing the
inode table during mkfs time fixes the problem.

Reproduction:
    1)  Start a GCE VM using Debian Stretch with an extra 10 GB PD-SSD
        disk attached to the VM.
    2)  Run the command "/sbin/mkfs.ext4 -I 2048 -b 4096 /dev/sdb 65536"
        followed by "mount /dev/sdb /mnt"
    3)  The system will be hung and after the SCSI timeout, SCSI tag
        abort messages will show up in the console.

Also disable lazy_journal_init while we're at it, since this could
cause potential failures in the shutdown tests if there is some
unexpected leftovers in the journal.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>