gce-xfstests: keep /results on boot disk, not separate results disk.
The entire /results directory was on a separately created disk, which
would not auto-delete. The purpose of this was so that if a VM wedged,
the results directory (whatever tests HAD finished) would be preserved
after the VM was shut off.
However, as boot disks can also be preserved, it's simple to not have a
/results disk at all, and not auto-delete the boot disk except in
the case of a graceful shutdown (when the tests finish), or if the user
forces it to be deleted (gce-xfstests abort). A normal shutdown
(gce-xfstests rm) will not delete the boot disk.
This also saves a bit of disk space usage in the project.

Tested: Ran gce-xfstests smoke three times.
A completed test run deleted the boot disk properly and uploaded the
results
"gce-xfstests abort" deleted both instance and boot disk.
"gce-xfstests rm" deleted the instance, but not the boot disk. I
launched a new instance, attached and mounted the boot disk's first
partition into an arbitrary directory, and was able to look into the
/results folder of the disk (as well as the rest of the files).

Signed-off-by: Jason Liu <liujason@google.com>
Change-Id: I1c4b501f6024e302cc86428bee383deaf9cea7db