gce-xfstests: identify username in test run name for get-results and ls
After the username in datecode change, get-results --unpack was
creating the directory "/tmp/results-".
This was due to the regex to match the test run ID only including
numbers, and not valid alphanumeric characters including hyphens,
which is now the case with test run IDs after the username in
datecode change.

Modified the regex to match alphanumeric and hyphens, and
"gce-xfstests get-results --unpack" now creates the correct temp
directory name.

Tested: ran get-results --unpack, directory created was
/tmp/results-USER-DATECODE instead of just /tmp/results-

Signed-off-by: Jason Liu <liujason@google.com>
Change-Id: Ia4c0a119ac02afc0bf524997922f05d8bdd5da2e