gce-xfstests: include gce-xfstests and python pkgs in create-image
First bit of multiple commits for the light test management server.
This commit modifies the creation of the test appliance image to include
a functional copy of gce-xfstests as well as some python packages.
The configuration files are piggybacked onto the GCS bucket config file,
and will be fetched by the LTM when it is booted up. One variable in
particular (GS_BUCKET) needs to be grabbed from metadata AFTER the
config is sourced, so the config source has been moved.

Signed-off-by: Jason Liu <liujason@google.com>
Change-Id: I53e67457879f160fcf52643d49a84e58b0d33ad2