gce-xfstests: sendgrid info in GCS, test appliances fetches on boot
The sendgrid API key/password/username used to be sent to the test
appliance in a metadata attribute, and deleted by the test appliance
after it was read. This involves sending this information from the
command line as a metadata attribute every time a new test appliance
is booted.

This change moves this to gce-do-setup, leveraging the fetch-cert
service (which has now been renamed to fetch-gs-files). The variables
are uploaded to the GCS bucket in a config file when setup is run.
The test appliance sources the config file at the start of gce-funcs.

Signed-off-by: Jason Liu <liujason@google.com>
Change-Id: I733831c2207cff82bcbb8b9516cff067ff5e87cf