gce-xfstests: summary file option to upload plaintext summary
A new option, GCE_UPLOAD_SUMMARY, has been added. This is uploaded to
the config stored in GCS. If the variable is nonzero, the shutdown
process will upload the /results/summary file to the GCS bucket in
plaintext (un-tar-xz'd)

This is for ease of viewing the results without having to use an installation
of gce-xfstests to extract the results (it's now easy to gsutil cat the
summary file to get an overview of the test run).

Tested: Built image. Added GCE_UPLOAD_SUMMARY to my own config, re-ran
the setup script. Config file was updated in GCS bucket.
Ran a single test, summary file was successfully uploaded.
Signed-off-by: Jason Liu <liujason@google.com>
Google-Bug-Id: 63674031
Change-Id: I095aa270cc165c6095b130c4c6adf5a63b8a2a01