gce-xfstests: command line override options for zone, bucket, run ID
Command line options added are --gce-zone, --gs-bucket, and --testrunid
In order to run a single test run in a separate zone, or use a
different GS bucket, it would require changing the
~/.config/gce-xfstests file. This change allows a single-time override
of the ZONE, GS_BUCKET, or the testrunid using 3 command line options.

The existing checks for a default configuration of GS_BUCKET and
GCE_ZONE have not been removed, and if a .config file based
configuration is not present, the added command line options will not
allow gce-xfstests to run.

--testrunid is preferred to be used over explicitly changing the
instance name via "--instance-name", as the test appliance tries to
parse out a testrunid by removing the "xfstests-" from its own instance
name, and if that doesn't match, the test appliance generates its own
testrunid using the date command, which is based on the datetime of the
test appliance.

Signed-off-by: Jason Liu <liujason@google.com>
Change-Id: Ia604efab1153342d68c34acc43c741660096fecd