gce-xfstests ltm: authenticate requests using 15 character password
As the /gce-xfstests POST endpoint will be used to launch child VMs and
is available to the public internet, the LTM needs to authenticate
requests coming to the endpoint. This change adds a login requirement
to the gce-xfstests POST endpoint.

gce-xfstests setup will generate a 15 character password using pwgen
and upload it to $GS_BUCKET/ltm-pass. Alternatively, the user can
manually upload the password.

The launch-ltm command will fail if the user does not have an
ltm-pass in their bucket. The launch-ltm command will also
download this password to a local hidden file.

Using the --ltm option on the command line will automatically
authenciate the request by logging in before making the request, so no
password needs to be manually entered.

Signed-off-by: Jason Liu <liujason@google.com>
Change-Id: I1d4b7e21501fe119593fd9c247dc129fca217af5