gce-xfstests ltm: launch a single child VM by parsing command.
When a command is received on the /gce-xfstests endpoint, a
TestRunManager object will be created.
The TestRunManager uses a Sharder to obtain Shards.
The Sharder currently isn't implemented, so it only uses the command
parser to parse the command line options for configs, and then
returns a single Shard containing the entire original test run.

The testrunmanager, when run, will launch itself into a new process,
and then launch shards into their own processes. Shards will launch
individual test VMs and monitor them. The testrunmanager process will
wait for the Shard processes to finish, before aggregating results
and exiting.

Signed-off-by: Jason Liu <liujason@google.com>
Change-Id: Ifd42e12fd858a87813190b6a67356135ee6879fa