gce-xfstests ltm: sharder splits shards by config, considers quota.
The sharder gets available project quotas for all regions in the
project, and determines the maximum number of shards it can create
based on the available quotas per region (2 CPUs and 1 IP address
required per test VM). Additionally, if a max_shards is specified, the
sharder will not exceed the given max, even if quotas would allow for
more.
The sharder will then create as many shards as is necessary and possible
for the configurations given. The zone that the shard will launch into
for each region is determined simply by picking the first available zone
seen in the region.
It is possible to add more complexity to this
process by passing more arguments from the user/configurations via
the options dictionary, and adding code to consider that.

Signed-off-by: Jason Liu <liujason@google.com>
Change-Id: Iaf6d825a31665b310843db1b8f6add90e2e960ed