gce-xfstests ltm: shard and testrunmanager aggregate results
Shard will now wait for test VM to complete. After the test VM completes,
the shard will look for a results file. The results will be unpacked
into a local dir before the shard exits, and if successful, the shard
will delete the results from the GCS bucket before exiting.

In waiting for the test VM, the shard wakes every 60 seconds to pull
serial port data from the test VM, and checks the status of the test
VM. If the status doesn't change for long enough, the shard
assumes the test VM has died and will move the serial port data into the
local dir instead of downloading the results file.

The testrunmanager, after all shards are complete, will find these local
unpacked results and aggregate them into a single results directory.
It will then re-tar and re-xz the directory, and upload it to the
bucket.

Signed-off-by: Jason Liu <liujason@google.com>
Change-Id: I3247041c0e4b31c5a6754a67ba6c73aff99f1de4