Robustify build rules for xfstests and xfsprogs
The makefile rules for xfstests and xfsprogs are wonderous and
strange.  Among other things, they will sometimes decide to call
configure of their own accord, which can cause the xfstests clean rule
to blow up since it runs configure without our special flags to tell
it where to find xfs/xfs.h.

This was happening if we updated xfsprogs or xfstests and the
makefiles thought (because of timestamp deltas) that it
needed to rebuild the Makefile and rerun the configure script.

To fix this, we need to set our configure options in
LOCAL_CONFIGURE_OPTIONS to pass them to the Makefile, and do the "make
clean" pass, if it is wanted, inside the build-all script.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>