gce-xfstests: fix cert generation logic in gce-do-setup
The test to decide whether or not gce-do-setup needs to generate a new
certificate was a bit broken.  First of all, when testing whether or
not a $REGEN_CERT is the null string, it must be quoted:

	test -z "$REGEN_CERT"

Secondly, we need to regenerate the certificate if either the private
key or certificate is missing.

Signed-off-by: Theodore Ts'o <tytso@google.com>
Change-Id: Ia63e486fff1a5fbd073e0b1be90f2ce7873bb1f0