Allow overriding chroot for do-all
Add a --chroot argument to do-all to allow easily setting the needed
BUILD_ENV and SUDO_ENV when building or updating a test appliance.  This
is useful to allow building both kvm-xfstests and android-xfstests test
appliances without having to edit config.custom, e.g:

   # build a root_fs.img for kvm-xfstests
   ./do-all --chroot=jessie-i386 --no-out-tar

   # build a root_fs.tar.gz for android-xfstests
   ./do-all --chroot=jessie-armhf --out-tar

As more complexity is being added to do-all, also take the opportunity
to add an --update option and make the do-update script just call
through to 'do-all --update'.  This makes it so that new logic doesn't
have to be duplicated in both scripts.

Signed-off-by: Eric Biggers <ebiggers@google.com>
Signed-off-by: Theodore Ts'o <tytso@mit.edu>