Build dynamically linked binaries by default
It seems the default of static linking was set several years ago, before
it was recommended to use a chroot to build the test appliances.  But
now it is pretty straightforward to set up a build chroot, and in that
case using static linking only bloats the binary sizes.  And although
static linking does make it a bit more plausible to build the
kvm-xfstests and gce-xfstests binaries outside a chroot, when actually
trying it I've ran into a number of problems, including:

    - distros shipping shared libraries but not static ones for some
      packages, causing the build to fail
    - the gen-image script making Debian-specific assumptions, such as
      assuming it can call apt-cache and dpkg, and running commands in
      the chroot without resetting the $PATH

Some of the problems can be solved, but not all can be, and dynamic
linking is preferable anyway.  So change the default to dynamic linking.
It's still possible to request static linking by setting EXEC_LDFLAGS
and EXEC_LLDFLAGS.

Signed-off-by: Eric Biggers <ebiggers@google.com>
Signed-off-by: Theodore Ts'o <tytso@mit.edu>