gce-xfstests: results and summaries in a results/ directory.
With a lot of test runs, the root directory of the GS bucket can get
very cluttered. In addition, the summary file change added a potential
second file to every single one of the test runs.
This change moves all results files to a results/ folder in the root of
the GS bucket. The LTM and the regular gce-xfstests scripts have been
updated to recognize this change.

Signed-off-by: Jason Liu <liujason@google.com>
Change-Id: I534b9e698f912ce39086e8bad2060d19423c1800