gce-xfstests: add bucket_subdir option to relocate results
This change is so that results could be located in different bucket
subdirectories, so that if a user is testing different versions of
a kernel, they could have the results uploaded to a subdir specific to
the kernel that was being tested instead of relying on the name of the
.tar.xz blob.

If the intention is to be testing a different kernel, of course, the
--kernel command line option will be necessary in order to not test just
the default configured GCE_KERNEL.

Setting a BUCKET_SUBDIR in a configuration and running setup will upload
the BUCKET_SUBDIR option as well to the GS bucket. By default, if no
--bucket-subdir override option is specified, all gce-xfstests runs
(including LTM runs) will use the BUCKET_SUBDIR option in the GS bucket.
If no BUCKET_SUBDIR is specified at all, the existing behavior will be
used. The results/ directory will be used for results, and the root of
the bucket will be used to hold the cached bzImage.

Signed-off-by: Jason Liu <liujason@google.com>
Change-Id: I30ab8e34503904837d18f0ac53ca25301b008479