gce-xfstests: add GCE_MIN_SCR_SIZE option
IOPS-bound tests take longer on GCE with smaller scratch disks because
the IOPS per disk is assigned based on the size of the disk (e.g.
generic/027).

This change adds a GCE_MIN_SCR_SIZE option, which should be an integer
value between 0 and 250 (gce-do-setup will validate this). If the
option is present, test appliances will compute the required size for
the tests that are being run, and if its less than GCE_MIN_SCR_SIZE, it
will instead create a disk with the size GCE_MIN_SCR_SIZE.

If GCE_MIN_SCR_SIZE is set to be greater than 56, it is essentially
guaranteed that the scratch disk size will be GCE_MIN_SCR_SIZE, as all
test and scratch partitions can fit into a 56 GB disk.

Signed-off-by: Jason Liu <liujason@google.com>
Change-Id: I329a3d6526ce31ec4874ef55c0cd76038653c6bf
Google-Bug-Id: 64518176